+++ Kernel/GenericInterface/Provider.pm
@@ -23,6 +23,7 @@ use Kernel::System::VariableCheck (qw(IsHashRefWithData));
 our @ObjectDependencies = (
     'Kernel::System::Log',
     'Kernel::System::GenericInterface::Webservice',
+    'Kernel::System::Valid',
 );
 
 =head1 NAME
@@ -115,6 +116,17 @@ sub Run {
             Priority => 'error',
             Message =>
                 "Could not load web service configuration for web service at $RequestURI",
+        );
+
+        return;    # bail out without Transport, Apache will generate 500 Error
+    }
+
+    # Check if web service has valid state.
+    if ( $Kernel::OM->Get('Kernel::System::Valid')->ValidLookup( ValidID => $Webservice->{ValidID} ) ne 'valid' ) {
+        $Kernel::OM->Get('Kernel::System::Log')->Log(
+            Priority => 'error',
+            Message =>
+                "Web service '$Webservice->{Name}' is not valid and can not be loaded",
         );
 
         return;    # bail out without Transport, Apache will generate 500 Error