+++ CHANGES.md
@@ -1,4 +1,5 @@
 #5.0.28 ????-??-??
+ - 2018-05-07 Fixed bug#[13831](https://bugs.otrs.org/show_bug.cgi?id=13831) - Invalid Web services are still working.
  - 2018-04-24 Fixed bug#[13805](https://bugs.otrs.org/show_bug.cgi?id=13805) - Distribution OpenBSD is not recognized by support data collector.
  - 2018-04-18 Fixed bug#[11132](https://bugs.otrs.org/show_bug.cgi?id=11132) - Loss of attached files with long Cyrillic names.
  - 2018-04-18 Fixed bug#[13554](https://bugs.otrs.org/show_bug.cgi?id=13554) - Non-allowed characters in attachment file names.